requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
google-auth
google-auth-oauthlib
google-api-python-client
PyMuPDF
python-docx
transformers
torch --index-url https://download.pytorch.org/whl/cpu
groq
fpdf2
pandas
python-dotenv
sentencepiece
